6. Tạo 2 model User và Person có quan hệ OnetoOne - Model Person bao gồm: full_name, gender, birthdate, phone_number, address - Model User bao gồm: email, password, is_active 7. CRUD cho User và Person

+++ app/Repositories/PersonRepository.php
@@ -10,24 +10,35 @@ class PersonRepository
     {
         return Person::all();
     }
+    public function getAllPersonsWithUser()
+    {
+        return Person::with('user')->get();
+    }
+    public function find($id)
+    {
+        return Person::find($id);
+    }
 
     public function create(array $data)
     {
         return Person::create($data);
     }
 
-    public function find($id)
+    public function update($id, array $data)
     {
-        return Person::find($id);
+        $person = $this->find($id);
+        if ($person) {
+            return $person->update($data);
+        }
+        return false;
     }
 
-    public function update(Person $person, array $data)
+    public function delete($id)
     {
-        return $person->update($data);
-    }
-
-    public function delete(Person $person)
-    {
-        return $person->delete();
+        $person = $this->find($id);
+        if ($person) {
+            return $person->delete();
+        }
+        return false;
     }
 }
